refactor: Clean up packages for new classes
