+++ common/src/main/java/net/blay09/mods/balm/network/internal/NetworkVersions.java
@@ -1,4 +1,4 @@
-package net.blay09.mods.balm.platform.module.internal;
+package net.blay09.mods.balm.network.internal;
 
 public record NetworkVersions(String modVersion, String networkVersion, boolean requireRemote) {
 }
